statable: Run Stata Commands from R

An R to Stata interface that allows Stata commands to be executed from within R. Data sets can easily be transferred between R and Stata, and Stata estimation results can be returned to R.

See also

Author

Maintainer: James Atkins james@jamesatkins.com